+++ app/src/components/postcard/Postcard.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import './postcard.css';
 
 type PostcardProps = {
@@ -13,11 +12,11 @@ type PostcardProps = {
 }
 
 
-function PostCard ({titulo, texto, isFilled, onClick}:PostcardProps){
+function PostCard({ titulo, texto, isFilled, onClick }: PostcardProps) {
 
   var image;
 
-  if (isFilled===true) {
+  if (isFilled === true) {
     image = "./images/liked.png";
   }
   else {
@@ -26,12 +25,12 @@ function PostCard ({titulo, texto, isFilled, onClick}:PostcardProps){
 
 
 
-  return(
+  return (
     <div className="postcard">
       <h3 className="titulo">{titulo}</h3>
       <p className="texto">{texto}</p>
-      <div className="cardbotton">     
-        <img className="like" src={image} onClick={onClick} alt="Like Button"/>
+      <div className="cardbotton">
+        <img className="like" src={image} onClick={onClick} alt="Like Button" />
       </div>
     </div>
   )
