+++ app/src/routes.tsx
@@ -1,0 +1,18 @@
+import {
+    Route,
+    Switch,
+    BrowserRouter as Router,
+} from 'react-router-dom';
+import Home from './pages/Home'
+
+const Routes = () => {
+    return (
+        <Router>
+            <Switch>
+                <Route component={Home} path='/' exact />
+            </Switch>
+        </Router>
+    );
+}
+
+export default Routes;
